+++ src/__init__.py
@@ -1,4 +1,4 @@
 """Fexobooth - Photobooth Software für fexobox"""
 
-__version__ = "2.2.4"
+__version__ = "2.2.5"
 __author__ = "fexon e.K."